Remove checks/sets for irbtmu_immobile_unit tag
GetTags() is an expensive operation during AI thinking/Combat as it must create and update a new TagSet at least once.

To speed up AI calcs IRBT was changed to no longer check for the irbtmu_immobile_unit tag on units during movement calculations and instead now solely relies on the stat version.

CU made checks for this tag when getting a Mechs Jump Distance and CAC used it to immobilize units when they perform artillery attacks. These checks and sets of the tag have been updated to use new helper methods in IRBTModUtils that check or set the stat instead of using the tag, fixing them for the updated IRBT and increasing performance by removing tagset based checks at the same time.