perf improvements to GetChassisTagsModifyer
found that GetChassisTagsModifyer was showing up as an issue in our performance profiling due to its heavy usage of tagset creation.

to fix this calculate a base value for a combantant, weapon, mode and ammo once on what should be static unit tags, then we only need to lookup this value and iterate over the encounter tags for the target to get the final modifier.

cached lookup values are cleared when contract Complete is called (at the end of combat) to prevent the cache from keeping stale references around