some internal methods of CC that CU was trying to access have been removed or changed. In order to use newer versions of CC, CU needed to be updated.
CU shouldn't really be accessing the internals of CC like this. to encapsulate this, I have moved all impacted methods/fields into a new CC Access Helper and used reflection, removing the publicizer from running on CC.

This may need to be further enhanced as CC continues to be worked on.

Ideally in the future, the vehicle lab is removed from CU and implemented into CC & ME where it probably belongs.